Fixes to stepping procedure, added checks, minor cleanup
Steps should now be correctly scaled to the sine array. Checks were added to MIN and MAX_MICROSTEP_DIVISOR to prevent user errors. Microstep divisors are now uint8_ts for space conservation (no need for uint16_t)

+++ src/config/config.h
@@ -154,8 +154,8 @@
 // The min/max microstepping divisors
 // Microstepping divisors are the numbers underneath the fraction of the microstepping
 // For example, 1/16th microstepping would have a divisor of 16
-#define MIN_MICROSTEP_DIVISOR   (uint8_t)1
-#define MAX_MICROSTEP_DIVISOR   (uint8_t)32
+#define MIN_MICROSTEP_DIVISOR   1
+#define MAX_MICROSTEP_DIVISOR   32
 
 // The mode to set the motor to when it's disabled
 #define IDLE_MODE               COAST
